+++ Assets/Materials/Tree.mat
@@ -105,7 +105,7 @@ Material:
     - _SpecularIntensity: 0.2
     - _SrcBlend: 1
     - _UVSec: 0
-    - _WindSpeed: 0.5
+    - _WindSpeed: 0.025
     - _WindStrength: 0.1
     - _ZWrite: 1
     - __dirty: 1
