+++ Assets/Prefabs/Grass Renderer.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 33a391a806c20fc47a4cac911454205f
-folderAsset: yes
-DefaultImporter:
+guid: 3472e3297a383e94ca044f6f040c7ca1
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
