+++ Assets/Scenes/Main/Main Post Process Profile.asset
@@ -1430,7 +1430,7 @@ MonoBehaviour:
     value: {r: 0, g: 0, b: 0, a: 1}
   ambientOnly:
     overrideState: 1
-    value: 0
+    value: 1
   noiseFilterTolerance:
     overrideState: 0
     value: 0
