+++ Assets/Scripts/Environment/Grass.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 98c2570aedec31c40ab7d046e92d6ed6
-NativeFormatImporter:
+guid: cb5a905958d782548b4d6574197e76bf
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 8600000
   userData: 
   assetBundleName: 
   assetBundleVariant:
