+++ Assets/Scripts/Environment/Grass/GrassInstance.compute.meta
@@ -1,9 +1,8 @@
 fileFormatVersion: 2
-guid: f5bddbb6213bed54a9ab86d1ea2854ac
-ShaderImporter:
+guid: f5854325310d6cc42b5f6ed5c8e27977
+ComputeShaderImporter:
   externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
+  currentAPIMask: 4
   preprocessorOverride: 0
   userData: 
   assetBundleName:
