+++ Assets/Scripts/Environment/Grass/Instanced_InstancedSurfaceShader.mat
@@ -8,23 +8,23 @@ Material:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: Instanced_InstancedSurfaceShader
-  m_Shader: {fileID: 4800000, guid: bdd74991912fa264baba247bc8dd5992, type: 3}
+  m_Shader: {fileID: 4800000, guid: 1b399f5cb349edd478149c263edfa3b5, type: 3}
   m_ShaderKeywords: 
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
-  m_CustomRenderQueue: 2000
+  m_CustomRenderQueue: -1
   stringTagMap: {}
   disabledShaderPasses: []
   m_SavedProperties:
     serializedVersion: 3
     m_TexEnvs:
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: 9ed2e9fdca976d44ea97ad0d79f93d7f, type: 3}
+        m_Texture: {fileID: 2800000, guid: 1b5badbc036582448be042d3ec9e2e0c, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MaskTex:
-        m_Texture: {fileID: 8600000, guid: 98c2570aedec31c40ab7d046e92d6ed6, type: 2}
+        m_Texture: {fileID: 8600000, guid: 2edc3d2701a48c2418bd626f6f7cae4b, type: 2}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0.5, y: 0.5}
     m_Floats:
@@ -35,10 +35,10 @@ Material:
     - _MaskSub: 0.65
     - _Metallic: 0
     - _Multiply: 252.3
-    - _RTSize: 25.6
+    - _RTSize: 22
     - _WindSpeed: 0.1
     - _WindStrength: 0.1
     m_Colors:
     - _Color: {r: 0.408, g: 0.51, b: 0.10199999, a: 1}
-    - _TipColor: {r: 0.6, g: 0.75, b: 0.14999999, a: 1}
+    - _TipColor: {r: 0.3, g: 0.3, b: 0.3, a: 1}
   m_BuildTextureStacks: []
