+++ Assets/Scripts/Environment/Grass/New Custom Render Texture.asset
@@ -17,8 +17,8 @@ CustomRenderTexture:
   m_Width: 128
   m_Height: 128
   m_AntiAliasing: 8
-  m_MipCount: -1
-  m_DepthFormat: 0
+  m_MipCount: 8
+  m_DepthFormat: 2
   m_ColorFormat: 49
   m_MipMap: 0
   m_GenerateMips: 1
@@ -36,7 +36,7 @@ CustomRenderTexture:
     m_WrapW: 1
   m_Dimension: 2
   m_VolumeDepth: 1
-  m_Material: {fileID: 2100000, guid: cb997a1495152194cbf703854bbc7c05, type: 2}
+  m_Material: {fileID: 0}
   m_InitSource: 0
   m_InitMaterial: {fileID: 0}
   m_InitColor: {r: 1, g: 1, b: 1, a: 1}
