+++ Assets/Scripts/Environment/Grass/Unlit_RTShader.mat
@@ -8,7 +8,7 @@ Material:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: Unlit_RTShader
-  m_Shader: {fileID: 4800000, guid: f5bddbb6213bed54a9ab86d1ea2854ac, type: 3}
+  m_Shader: {fileID: 4800000, guid: dc11108df9826a1458afe905951178d7, type: 3}
   m_ShaderKeywords: 
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
@@ -20,7 +20,7 @@ Material:
     serializedVersion: 3
     m_TexEnvs:
     - _Tex:
-        m_Texture: {fileID: 8600000, guid: 98c2570aedec31c40ab7d046e92d6ed6, type: 2}
+        m_Texture: {fileID: 8600000, guid: 2edc3d2701a48c2418bd626f6f7cae4b, type: 2}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats: []
