+++ Assets/Scripts/Environment/Lakes.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: cb997a1495152194cbf703854bbc7c05
-NativeFormatImporter:
+guid: 982e6b37cded53f47ab2a0b7cdda3249
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
