+++ ProjectSettings/ProjectSettings.asset
@@ -145,7 +145,8 @@ PlayerSettings:
   resolutionScalingMode: 0
   androidSupportedAspectRatio: 1
   androidMaxAspectRatio: 2.1
-  applicationIdentifier: {}
+  applicationIdentifier:
+    Standalone: com.TeamALLCAPS.Ozymandias
   buildNumber:
     Standalone: 0
     iPhone: 0
